healthsync/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery[redis]
redis
django-environ
django-cors-headers
drf-spectacular
qrcode
 Pillow
